refactor: Section을 line package로 이동
