+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -15,8 +15,8 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
+import nextstep.subway.line.domain.Section;
 import nextstep.subway.line.dto.LineRequest;
-import nextstep.subway.section.domain.Section;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.domain.Station;
