+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -16,8 +16,8 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.LineAcceptanceTest;
+import nextstep.subway.line.domain.Section;
 import nextstep.subway.line.dto.AddSectionRequest;
-import nextstep.subway.section.domain.Section;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.domain.Station;
